% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.CMIP5.R
\name{CMIP5.files.for.timeslice}
\alias{CMIP5.files.for.timeslice}
\title{Test file names against desired date range}
\usage{
CMIP5.files.for.timeslice(candidates, start, end)
}
\arguments{
\item{candidates}{vector of character file names with embedded date strings}

\item{start}{character string giving period start time 'YYYY-MM-DD:HH:MM'}

\item{end}{character string giving period end time 'YYYY-MM-DD:HH:MM'}
}
\value{
vector of TRUE/False - is target time in file.
}
\description{
Variables are often split into files by date, with the file name
 marking the date range.
}
\details{
This function tests file names against a specified date range, to find
 those that contain data covering those dates.
}
