% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{CMIP5.get.climatology.file.name}
\alias{CMIP5.get.climatology.file.name}
\title{CMIP5 get file name (hourly climatologies)}
\usage{
CMIP5.get.climatology.file.name(model = NULL, experiment = NULL,
  variable = NULL, ensemble = "r1i1p1", realm = "atmos", table = "3hr",
  frequency = "3hr")
}
\arguments{
\item{model}{- 'HadGEM2-A','CCSM4'  - or any model name}

\item{experiment}{- 'historical', 'rcp85', '1pctCO2' etc..}

\item{variable}{- 'tas', 'pr', 'uas', 'vas' - or any short name}

\item{ensemble}{-  r<N>i<M>p<L> - defaults to 'r1i1p1'.}

\item{realm}{- 'atmos' (default), 'ocean', 'land', etc.}

\item{table}{- see http://cmip-pcmdi.llnl.gov/cmip5/docs/standard_output.pdf, defaults to '3hr'.}

\item{frequency}{'3hr' (default), 'mon', 'yr', 'fx', etc.}
}
\value{
File names for the climatology data
}
\description{
Get the file name for selected climatological data.
}
\details{
Called internally by \code{CMIP5.make.climatology.at.hour} but also useful
 called directly - you can then access the data with another tool.
The vocbulary for the options is given in
 http://cmip-pcmdi.llnl.gov/cmip5/docs/cmip5_data_reference_syntax.pdf
}
