% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.CMIP5.R
\name{CMIP5.get.data.dir}
\alias{CMIP5.get.data.dir}
\title{CMIP5 get data directory}
\usage{
CMIP5.get.data.dir(model = NULL, experiment = NULL, variable = NULL,
  ensemble = NULL, realm = NULL, table = NULL, frequency = NULL,
  year = NULL, month = NULL, day = NULL, hour = NULL)
}
\value{
Base directory name vector - NULL for no possibilities.
}
\description{
Find directories that may have data in, there may be several
}
\details{
Data may be in any of several different directory trees.
This function returns a list of possible base directories.
}
