% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.CMIP5.R
\name{CMIP5.get.institute.from.model}
\alias{CMIP5.get.institute.from.model}
\title{CMIP5 get Institute from model}
\usage{
CMIP5.get.institute.from.model(model)
}
\arguments{
\item{model}{CMIP5 model name}
}
\value{
Name of institute behind model.
}
\description{
It's redundant to specify the institude, just the model is sufficient.
But we need the institute to find the data - get it from the model.
}
\details{
Requires a table relating institutes and models but this doesn't
 change often.
}
