% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.CMIP5.R
\name{CMIP5.get.slab.climatology}
\alias{CMIP5.get.slab.climatology}
\title{Extract a hyperslab of climatology}
\usage{
CMIP5.get.slab.climatology(model = NULL, experiment = NULL,
  variable = NULL, ensemble = "r1i1p1", realm = "atmos", table = "3hr",
  frequency = "3hr", date.range, calendar = "gregorian",
  height.range = NULL, lat.range = c(-90, 90), lon.range = c(-180, 360))
}
\arguments{
\item{model}{- 'HadGEM2-A','CCSM4'  - or any model name}

\item{experiment}{- 'historical', 'rcp85', '1pctCO2' etc..}

\item{variable}{- 'tas', 'pr', 'uas', 'vas' - or any short name}

\item{ensemble}{-  r<N>i<M>p<L> - defaults to 'r1i1p1'.}

\item{realm}{- 'atmos' (default), 'ocean', 'land', etc.}

\item{table}{- see http://cmip-pcmdi.llnl.gov/cmip5/docs/standard_output.pdf, defaults to '3hr'.}

\item{frequency}{'3hr' (default), 'mon', 'yr', 'fx', etc.}

\item{date.range}{A pair of date strings, e.g. c('1981-02-05:12','1981-03-27:18') - inclusive.}

\item{calendar}{Calendar used by data - 'gregorian' (default), '360_day' or '365_day'.}

\item{height.range}{Bottom and top heights in hPa - leave NULL for monolevel}

\item{lat.range}{Min and max latitude in degrees - defaults to c(-90,90)}

\item{lon.range}{Min and max longitude in degrees - defaults to c(0,360)}

\item{variable}{20CR variable name, only 2d variables will work.}
}
\value{
A GSDF field with the selected multidimensional data
}
\description{
Up to 4d (lat, lon, height, time).
}
\details{
Don't call this directly - called from CMIP5.get.slab
}
