% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{CMIP5.make.climatology}
\alias{CMIP5.make.climatology}
\title{Make (and store) a climatological average}
\usage{
CMIP5.make.climatology(model = NULL, experiment = NULL, variable = NULL,
  ensemble = "r1i1p1", realm = "atmos", table = "3hr",
  frequency = "3hr", height.range = NULL, lat.range = c(-90, 90),
  lon.range = c(-180, 360), calendar = "gregorian", first.year, last.year)
}
\arguments{
\item{model}{- 'HadGEM2-A','CCSM4'  - or any model name}

\item{experiment}{- 'historical', 'rcp85', '1pctCO2' etc..}

\item{variable}{- 'tas', 'pr', 'uas', 'vas' - or any short name}

\item{ensemble}{-  r<N>i<M>p<L> - defaults to 'r1i1p1'.}

\item{realm}{- 'atmos' (default), 'ocean', 'land', etc.}

\item{table}{- see http://cmip-pcmdi.llnl.gov/cmip5/docs/standard_output.pdf, defaults to '3hr'.}

\item{frequency}{'3hr' (default), 'mon', 'yr', 'fx', etc.}

\item{calendar}{- calendar used by the data 'gregorian', '360_day', or '365_day'}

\item{first.year}{first year of climatological period}

\item{last.year}{last year of climatological period}

\item{variable}{short variable name 'tas', 'pr', uas, 'ps' etc.}
}
\value{
Nothing - a climatological field will be created as a side effect.
}
\description{
Data are stored as .nc files in a parallel directory to the original
 data - same as original data exept only for 1 year.
Loads a whole year of data at once - high resolution 3D fields
 may run into memory problems.
}
\details{
Chose to use 1981 as the storage year in the netcdf file.
}
